+++ CHANGELOG.md
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+### [0.2.0-beta] - 2024-10-22
+
+### Feat
+
+- feat: use toJSON method from axios instance
+- fix: ensure delete logs files for 0 days
+
+---
+
+## Unreleased
+
 ### [0.1.5] - 2023-12-15
 
 ### Fix